+++ kameleoon/client/src/Events/DataFileUpdateHandler.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kameleoon\Events;
+
+/**
+ * Handler of SDK data file update events.
+ */
+interface DataFileUpdateHandler extends EventHandler
+{
+    /**
+     * Called when the SDK configuration (data file) is updated.
+     */
+    public function onUpdate(DataFileUpdateEvent $event): void;
+}