+++ kameleoon/client/src/Events/EventHandler.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kameleoon\Events;
+
+/**
+ * Base interface for all SDK event handlers.
+ */
+interface EventHandler
+{
+}