+++ kameleoon/client/src/Exception/KameleoonException.php
@@ -3,18 +3,19 @@
 namespace Kameleoon\Exception;
 
 use Exception;
+use Throwable;
 
 class KameleoonException extends Exception
 {
     private const KAMELEOON_SDK = "Kameleoon SDK: ";
 
-    public function __construct($message, $code = 0, ?Exception $previous = null)
+    public function __construct($message, ?Throwable $previous = null)
     {
-        parent::__construct(self::KAMELEOON_SDK . $message, $code, $previous);
+        parent::__construct(self::KAMELEOON_SDK . $message, 0, $previous);
     }
 
     public function __toString()
     {
-        return __CLASS__ . ": [{$this->code}]: {$this->message}\n";
+        return get_class($this) . ": {$this->message}\n";
     }
 }