+++ kameleoon/client/src/Network/NetworkManagerFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kameleoon\Network;
 
+use Kameleoon\Events\EventManager;
 use Kameleoon\Network\AccessToken\AccessTokenSourceFactory;
 
 interface NetworkManagerFactory
@@ -14,6 +15,7 @@ interface NetworkManagerFactory
         int $defaultTimeout,
         string $kameleoonWorkDir,
         AccessTokenSourceFactory $accessTokenSourceFactory,
-        ?string $networkDomain
+        ?string $networkDomain,
+        EventManager $eventManager
     ): NetworkManager;
 }